% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaderfunctions.R
\name{AddFixGroups}
\alias{AddFixGroups}
\title{Yhdistä ryhmiä, jotka on pitänyt hakea kahtena erillisenä
ryhmänä korpuksista}
\usage{
AddFixGroups(filtered)
}
\description{
Yhdistä ryhmiä, jotka on pitänyt hakea kahtena erillisenä
ryhmänä korpuksista
}
