% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving_data.R
\name{AddFolder}
\alias{AddFolder}
\title{Lis채채 kokonaisen kansiollisen json-tiedostoja dataan}
\usage{
AddFolder(path, thislang)
}
\arguments{
\item{path}{polku kansioon}

\item{thislang}{kieli}
}
\description{
Lis채채 kokonaisen kansiollisen json-tiedostoja dataan
}
