% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving_data.R
\name{AddGroupFromJson}
\alias{AddGroupFromJson}
\title{Lisää uuden ryhmän suoraan json-tiedostosta}
\usage{
AddGroupFromJson(path, lang)
}
\arguments{
\item{path}{plku jsontiedostoon}

\item{lang}{kieli}
}
\description{
Lisää uuden ryhmän suoraan json-tiedostosta
}
