% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{BuildAccuracyTest}
\alias{BuildAccuracyTest}
\title{Haetaan esimerkkejä parserin tarkkuuden määrittelemiseksi}
\usage{
BuildAccuracyTest(alldata, sample_size)
}
\arguments{
\item{alldata}{data, josta lähdetään liikkeelle (df)}

\item{sample_size}{kuinka monta lausetta kustakin sijainnista halutaan}
}
\value{
tibble, jossa testit
}
\description{
Haetaan esimerkkejä parserin tarkkuuden määrittelemiseksi
}
