% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{CheckSample_df}
\alias{CheckSample_df}
\title{Kysyy määritelmää jollekin esimerkille tms.}
\usage{
CheckSample_df(r, cols_to_show,
  backup_file = "/home/juho/drive/backups/backup_for_checksample.txt")
}
\arguments{
\item{r}{yksi dataframen rivi}

\item{cols_to_show}{vektori, jossa on niitten sarakkeiden nimet, joita halutaan näyttää päätöksenteon pohjaksi}

\item{backup_file}{polku tiedostoon, jonne jokainen vastaus tallennetaan varmuuskopiona}
}
\value{
Käyttäjän kirjoittama merkkijono
}
\description{
Käytetään yleensä apply-funktion callbackinä manuaalista annotointia tms. varten
}
