% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{CheckSample_simple}
\alias{CheckSample_simple}
\title{Kysyy määritelmää jollekin esimerkille tms.}
\usage{
CheckSample_simple(show_this,
  backup_file = "/home/juho/drive/backups/backup_for_checksample.txt")
}
\arguments{
\item{show_this}{elementti, joka käyttäjälle näytetään päätöksenteon pohjaksi}

\item{backup_file}{polku tiedostoon, jonne jokainen vastaus tallennetaan varmuuskopiona}
}
\value{
Käyttäjän kirjoittama merkkijono
}
\description{
Käytetään yleensä sapply-funktion callbackinä manuaalista annotointia tms. varten
}
