% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collocations.R
\name{CollocateTable}
\alias{CollocateTable}
\title{Luo SimpleCollocations-funktion avulla taulukon eri kollokaattien yleisyydestä
Lisäksi tallentaa yksittäisten kollokaattien tiedot ja palauttaa kaiken informaation yhtenä isona listana}
\usage{
CollocateTable(mydf, colpat, colrange = 3)
}
\arguments{
\item{mydf}{data.frame, jonka pohjalta luodaan}

\item{colpat}{Regex siitä, minkä kollokaatteja etsitään}
}
\description{
Luo SimpleCollocations-funktion avulla taulukon eri kollokaattien yleisyydestä
Lisäksi tallentaa yksittäisten kollokaattien tiedot ja palauttaa kaiken informaation yhtenä isona listana
}
