% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d}
\alias{d}
\title{Tutkimuksen tilastollisen mallin ym. kannalta oleellinen numeerinen data}
\format{tibble seuraavalla rakenteella
\describe{
  \item{lang}{lauseen kieli}
  \item{group}{esimerkin aineistoryhmä}
  \item{funct}{semanttinen funktio}
  \item{morph}{morfologinen rakenne}
  \item{location3}{kolmijakoinen sijaintimuuttuja (S1, S2/S3, S4)}
  \item{location}{nelijakoinen sijaintimuuttuja (S1,S2,S3,S4)}
  \item{isnumeric}{onko lauseessa läsnä numeraali - tarkemmin ks. funktio DoesThisRowHaveNumeral}
  \item{pos}{positionaalisuus}
  \item{red}{referentiaalisuus}
  \item{corpustype}{lähdekorpuksen tyyppi}
  \item{sentid}{lauseen alkuperäinen id tietokannassa}
  ...
}}
\usage{
d
}
\description{
Tästä julkisesti saatavilla olevasta datasta tarkoituksella jätetty pois mahdollisesti sensitiivinen, varsinaisia tekstejä sisältävä korpusdata
}
\keyword{datasets}
