% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaderfunctions.R
\name{DefineGroupProperties}
\alias{DefineGroupProperties}
\title{Apufunktio semanttisen funktion ym. m채채rittelyyn}
\usage{
DefineGroupProperties(name, propname, conv.table)
}
\description{
Apufunktio semanttisen funktion ym. m채채rittelyyn
}
