% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurelists.R
\name{DetermineAdvPhraseLength}
\alias{DetermineAdvPhraseLength}
\title{Määrittele, kuinka pitkä on ajanilmauksen sisältämä lauseke}
\usage{
DetermineAdvPhraseLength(thisrow, feats_vector = c())
}
\value{
lista, jonka arvoina lausekkeen pituus molempiin suuntiin (substractedmargin=x,addedmargin=x)
}
\description{
Määrittele, kuinka pitkä on ajanilmauksen sisältämä lauseke
}
