% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getting_numeric_cases.R
\name{DoesThisRowHaveNumeral}
\alias{DoesThisRowHaveNumeral}
\title{Määrittele jostakin tutkimusdatataulukon rivistä, onko siinä (ajanilmauksen lisäksi) jokin numeraali}
\usage{
DoesThisRowHaveNumeral(thisrow)
}
\description{
Numeraalin määrittelylle asetetaan seuraavat kriteerit: 1.kardinaaliluku 2. ei tyypillinen vuosiluku 3. ei ilmaus n vuotta
LISÄKSI oletetaan, että seuraavan sanan on oltava suomessa partitiivissa, venäjässä genetiivissä.
}
