% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_getters.R
\name{FilterPronouns}
\alias{FilterPronouns}
\title{Palauttaa listan, jossa inputtina annettu dataframe on subsetattu sen mukaan, mitä pronomineja siinä on käyttäjän määrittelemässä roolissa (esim. subjektina)}
\usage{
FilterPronouns(d = data.frame(), filterby = character())
}
\arguments{
\item{d}{dataframe, jota filtteröidään}

\item{filterby}{sen sarakkeen nimi, josta pronomineja etsitään}
}
\description{
Palauttaa listan, jossa inputtina annettu dataframe on subsetattu sen mukaan, mitä pronomineja siinä on käyttäjän määrittelemässä roolissa (esim. subjektina)
}
