% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaderfunctions.R
\name{FixGroupsFi}
\alias{FixGroupsFi}
\title{Tee j채lkikorjauksia aineistoryhmiin esimerkiksi poistamalla
tapauksia, joita aineistossa ei pit채isi olla}
\usage{
FixGroupsFi(filtered)
}
\description{
Tee j채lkikorjauksia aineistoryhmiin esimerkiksi poistamalla
tapauksia, joita aineistossa ei pit채isi olla
}
