% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaderfunctions.R
\name{FixGroupsRu}
\alias{FixGroupsRu}
\title{Tee j채lkikorjauksia aineistoryhmiin esimerkiksi poistamalla
tapauksia, joita aineistossa ei pit채isi olla}
\usage{
FixGroupsRu(filtered)
}
\description{
Tee j채lkikorjauksia aineistoryhmiin esimerkiksi poistamalla
tapauksia, joita aineistossa ei pit채isi olla
}
