% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{FormatExampleSource}
\alias{FormatExampleSource}
\title{Muotoile esimerkin lähde raakadatasta luettavampaan muotoon}
\usage{
FormatExampleSource(exrow)
}
\arguments{
\item{exrow}{dataframen rivi, jossa lause ja lähde}
}
\description{
Muotoile esimerkin lähde raakadatasta luettavampaan muotoon
}
