% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurelists.R
\name{FormatFeatString}
\alias{FormatFeatString}
\title{muokkaa kaikkien lauseen sanojen morfologisen rakenteen sisältävän merkkijonon tietoja
HUOM! Tämä on spesifisti ajatellen kysymystä siitä, seuraako ajanilmausta jokin numeraali!}
\usage{
FormatFeatString(thisrow)
}
\description{
muokkaa kaikkien lauseen sanojen morfologisen rakenteen sisältävän merkkijonon tietoja
HUOM! Tämä on spesifisti ajatellen kysymystä siitä, seuraako ajanilmausta jokin numeraali!
}
