% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_getters.R
\name{GetAdvShare}
\alias{GetAdvShare}
\title{Tähän tiedostoon on listattu funktioita, joiden tehtävä on, mahdollisesti
aika kertakäyttöisesti, hakea dataa tietyssä muodossa tai tietyllä tavalla
filtteröitynä isosta d-dataframesta
Hae adverbien osuus tietyssä sijainnissa}
\usage{
GetAdvShare(x, lang, total, location)
}
\arguments{
\item{lang}{kieli}

\item{total}{dataframe, josta haetaan}

\item{location}{sijainti, josta haetaan}
}
\description{
Tähän tiedostoon on listattu funktioita, joiden tehtävä on, mahdollisesti
aika kertakäyttöisesti, hakea dataa tietyssä muodossa tai tietyllä tavalla
filtteröitynä isosta d-dataframesta
Hae adverbien osuus tietyssä sijainnissa
}
