% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving_data.R
\name{GetD}
\alias{GetD}
\title{Muodostaa karsitun, nopeasti viitattavan version tutkimusdatasta}
\usage{
GetD(rawdata)
}
\arguments{
\item{customdataset}{jos annetaan lähtökohtana oleva data frame argumenttina}
}
\description{
Muodostaa karsitun, nopeasti viitattavan version tutkimusdatasta
}
