% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_if_has_place.R
\name{GetFeatsInRelationToTe}
\alias{GetFeatsInRelationToTe}
\title{Hae jonkin suhteessa ajanilmaukseen määritellyn sanan morfologiset piirteet}
\usage{
GetFeatsInRelationToTe(r, tepos, direction)
}
\arguments{
\item{r}{rivi vaihe1.df-df:stä}

\item{tepos}{ajanilmauksen sijainti}

\item{direction}{kuinka mones sana ja mihin suuntaan (negatiivinen taaksepäin)}
}
\description{
Hae jonkin suhteessa ajanilmaukseen määritellyn sanan morfologiset piirteet
}
