% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupmeta.R
\name{GetGroupMetaTab}
\alias{GetGroupMetaTab}
\title{Tulostaa taulukon, jossa haluttujen aineistoryhmien ominaisuudet on listattu tarkemmin}
\usage{
GetGroupMetaTab(gnames)
}
\arguments{
\item{gnames}{tulostettavien aineistoryhmien nimet}

\item{caption}{taulukon otsikko}
}
\description{
Tulostaa taulukon, jossa haluttujen aineistoryhmien ominaisuudet on listattu tarkemmin
}
