% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving_data.R
\name{GetKehRes}
\alias{GetKehRes}
\title{Laske kehyksiseen resultatiiviseen funktioon liittyviä tilastoja erityisesti siitä, missä lauseissa on mukana numeraali}
\usage{
GetKehRes()
}
\description{
Laske kehyksiseen resultatiiviseen funktioon liittyviä tilastoja erityisesti siitä, missä lauseissa on mukana numeraali
}
