% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving_data.R
\name{GetL1aDataForS2S3}
\alias{GetL1aDataForS2S3}
\title{Hakee monia l1a-aineistoon ja keskisijaintiin liittyvi채 tietoja erityisesti koskien ns. julkaisu- / luomisverbej채}
\usage{
GetL1aDataForS2S3()
}
\description{
Hakee monia l1a-aineistoon ja keskisijaintiin liittyvi채 tietoja erityisesti koskien ns. julkaisu- / luomisverbej채
}
