% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_getters.R
\name{GetL2Collocates}
\alias{GetL2Collocates}
\title{Hae l2-aineistoryhmien kollokaatteja ja muodosta tarvittavat taulukot yms.}
\usage{
GetL2Collocates()
}
\description{
Hae l2-aineistoryhmien kollokaatteja ja muodosta tarvittavat taulukot yms.
}
