% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving_data.R
\name{GetNonFocalTest}
\alias{GetNonFocalTest}
\title{Pyrkii rajaamaan S4-aineistoa niin, että siitä erotellaan tapauksia, jotka
*ovat* fokaalisia tapauksista, jotka eivät ole.}
\usage{
GetNonFocalTest()
}
\description{
Pyrkii rajaamaan S4-aineistoa niin, että siitä erotellaan tapauksia, jotka
*ovat* fokaalisia tapauksista, jotka eivät ole.
}
