% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving_data.R
\name{GetNumericCasesInGroups}
\alias{GetNumericCasesInGroups}
\title{Hae jokaisesta aineistoryhm채st채 ne lauseet, joissa on numeraali}
\usage{
GetNumericCasesInGroups()
}
\description{
Hae jokaisesta aineistoryhm채st채 ne lauseet, joissa on numeraali
}
