% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_getters.R
\name{GetPublishVerbs}
\alias{GetPublishVerbs}
\title{Hae kaikki yhden kielen "julkaisuverbeiksi" luokitellut verbit pakettiin sisällytestä datatiedostosta}
\usage{
GetPublishVerbs(lang, groupname)
}
\arguments{
\item{lang}{kieli (fi/ru)}

\item{groupname}{aineistoryhmä, jonka perusteella verbejä etsitään}
}
\description{
Hae kaikki yhden kielen "julkaisuverbeiksi" luokitellut verbit pakettiin sisällytestä datatiedostosta
}
