% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{GetSamplesForGroup}
\alias{GetSamplesForGroup}
\title{Hakee joka ryhmästä n satunnaista esimerkkiä}
\usage{
GetSamplesForGroup(g, langsubset, sample_size)
}
\arguments{
\item{g}{aineistoryhmän nimi}

\item{langsubset}{yhden kielen kaikki lauseet (df)}

\item{sample_size}{kuinka monta lausetta kustakin sijainnista halutaan}
}
\value{
tästä ryhmästä haetut esimerkkilauseet sijainneittain järjestettynä listana
}
\description{
Hakee joka ryhmästä n satunnaista esimerkkiä
}
