% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{GetSamplesForLang}
\alias{GetSamplesForLang}
\title{Hakee joka kielestä jokaisen aineistoryhmän ja joka aineistoryhmästä n esimerkkiä}
\usage{
GetSamplesForLang(l, alldata, sample_size)
}
\arguments{
\item{l}{kielen nimi}

\item{alldata}{koko aineistoa edustava dataframe, josta lähdetään liikkeelle}

\item{sample_size}{kuinka monta lausetta kustakin sijainnista halutaan}
}
\description{
Hakee joka kielestä jokaisen aineistoryhmän ja joka aineistoryhmästä n esimerkkiä
}
