% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{GetSamplesForPos}
\alias{GetSamplesForPos}
\title{Hakee joka sijainnista n määrän esimerkkejä (jos löytyy)}
\usage{
GetSamplesForPos(thisloc, groupsubset, sample_size)
}
\arguments{
\item{thisloc}{sijainti (S1-S4)}

\item{groupsubset}{yhden kielen yhden aineistoryhmän kaikki lauseet (df)}

\item{sample_size}{kuinka monta lausetta kustakin sijainnista halutaan}
}
\value{
satunnaisotanta lauseita
}
\description{
Hakee joka sijainnista n määrän esimerkkejä (jos löytyy)
}
