% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_if_has_place.R
\name{GetTeNoWithoutPunct}
\alias{GetTeNoWithoutPunct}
\title{Laske, mones sana ajanilmaus on virkkeessä, kun välimerkkejä ei huomioida}
\usage{
GetTeNoWithoutPunct(r)
}
\arguments{
\item{r}{rivi vaihe1.df-df:stä}
}
\description{
Laske, mones sana ajanilmaus on virkkeessä, kun välimerkkejä ei huomioida
}
