% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{GiveExample}
\alias{GiveExample}
\title{Hae määritellystä osa-aineistoista viisi (tai niin paljon kuin löytyy) satunnaista esimerkkiä ja tulosta ne lähteineen}
\usage{
GiveExample(subsetted)
}
\arguments{
\item{subsetted}{tutkimusaineiston osa, josta haetaan}
}
\description{
Hae määritellystä osa-aineistoista viisi (tai niin paljon kuin löytyy) satunnaista esimerkkiä ja tulosta ne lähteineen
}
