% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getting_numeric_cases.R
\name{GiveNumericals}
\alias{GiveNumericals}
\title{Returns the ratio of sentences with a numeral compared to all sentences in a specific data group}
\usage{
GiveNumericals(gname, mydf, lang, loc = "S1")
}
\description{
Returns the ratio of sentences with a numeral compared to all sentences in a specific data group
}
