% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groupmeta}
\alias{groupmeta}
\title{Kunkin aineistoryhmän ominaisuudet taksonomioiden 1--3 valossa}
\format{tibble}
\usage{
groupmeta
}
\description{
Kunkin aineistoryhmän ominaisuudet taksonomioiden 1--3 valossa
}
\keyword{datasets}
