% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving_data.R
\name{GroupSubsets}
\alias{GroupSubsets}
\title{Luo koko tutkimusdatasta aineistoryhmäkohtaisia alijoukkoja.}
\usage{
GroupSubsets(groupnames)
}
\arguments{
\item{groupname}{lista ryhmistä, joista alijoukot luodaan}
}
\description{
Luo koko tutkimusdatasta aineistoryhmäkohtaisia alijoukkoja.
}
