% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_utilities.R
\name{listmembers}
\alias{listmembers}
\title{Tulostaa listan tiettyyn semanttiseen funktioon kuuluvista aineistoryhmistä}
\usage{
listmembers(group, returnvector = F)
}
\arguments{
\item{group}{ryhmän nimi}

\item{returnvector}{halutaanko, että funktio palauttaa vektorin ryhmistä}
}
\description{
Tulostaa listan tiettyyn semanttiseen funktioon kuuluvista aineistoryhmistä
}
