% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_utilities.R
\name{listrefmembers}
\alias{listrefmembers}
\title{Tulostaa listan tiettyyn referentiaalisuuden tyyppiin kuuluvista
aineistoryhmistä}
\usage{
listrefmembers(group)
}
\arguments{
\item{group}{ryhmän nimi}
}
\description{
Tulostaa listan tiettyyn referentiaalisuuden tyyppiin kuuluvista
aineistoryhmistä
}
