% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood.R
\name{loglikelihood.test}
\alias{loglikelihood.test}
\title{R function for calculating LL, by Andrew Hardie, Lancaster University.}
\usage{
loglikelihood.test(O)
}
\arguments{
\item{O}{matrix in the format 2 X 2 e.g. matrix(c(52,50000,57,75000),byrow=F,ncol=2)}
}
\description{
http://corpora.lancs.ac.uk/sigtest/process.php?action=seeR
}
\details{
CHeckout rfast package, too (requires libgls-dev)
}
