% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupmeta.R
\name{MakeGroupMetaTibble}
\alias{MakeGroupMetaTibble}
\title{Tallentaa aineistoryhmien taksonomioiden 1--3 mukaiset ominaisuudet tibbleksi}
\usage{
MakeGroupMetaTibble()
}
\description{
Tallentaa aineistoryhmien taksonomioiden 1--3 mukaiset ominaisuudet tibbleksi
}
