% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_if_has_place.R
\name{MarkIfHasPlace}
\alias{MarkIfHasPlace}
\title{Tarkastele, onko ajanilmauksen vieressä paikanilmaus
MUTTA rajaa niin, että vain isolla kirjaimella alkavat paikanilmaukset on otettu mukaan
TAI lievempänä niin, että jos kaksi sanaa sitten tai kahden sanan päästä alkaa isolla...}
\usage{
MarkIfHasPlace(r, direction)
}
\arguments{
\item{r}{dataframen rivi}
}
\description{
Tarkastele, onko ajanilmauksen vieressä paikanilmaus
MUTTA rajaa niin, että vain isolla kirjaimella alkavat paikanilmaukset on otettu mukaan
TAI lievempänä niin, että jos kaksi sanaa sitten tai kahden sanan päästä alkaa isolla...
}
