% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parser_accuracy_test_simple}
\alias{parser_accuracy_test_simple}
\title{Tutkimuksen varsinaiset kontekstit}
\format{tibble

"sents"
Esimerkkilauseet, joilla testataan parserin tarkkuutta

Aineistoksi filtteröidyistä lauseista valittiin satunnaisotannalla}
\usage{
parser_accuracy_test_simple
}
\description{
(data ei julkista)
}
\keyword{datasets}
