% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pragtab.R
\docType{class}
\name{PragmaticAssertionSentence-class}
\alias{PragmaticAssertionSentence-class}
\title{S4 class for easy representation of tables of Lambrechtian pragmatic assertions and presuppositions. This class represents sentences}
\description{
S4 class for easy representation of tables of Lambrechtian pragmatic assertions and presuppositions. This class represents sentences
}
