% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pragtab.R
\docType{class}
\name{PragmaticAssertionTab-class}
\alias{PragmaticAssertionTab-class}
\title{S4 class for easy representation of tables of Lambrechtian pragmatic assertions and presuppositions. This class represents the actual tables}
\description{
S4 class for easy representation of tables of Lambrechtian pragmatic assertions and presuppositions. This class represents the actual tables
}
