% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ordered_difplot.R
\name{PrepareOrderedDifPlot}
\alias{PrepareOrderedDifPlot}
\title{Tulostaa kuvion, jossa vertailtu suomen ja venäjän prosenttiosuuksia}
\usage{
PrepareOrderedDifPlot(subsetted, loc, dif.treshold, langs, percentual = T,
  means = vector())
}
\arguments{
\item{subsetted}{dataframe, josta kuvio tehdään}

\item{loc}{sijainti, jota tutkitaan}

\item{dif.treshold}{kuinka pienet erot jätetään pois kuviosta}

\item{langs}{vertailtavat kielet}

\item{percentual}{järjestetäänkö suhteellisen eron mukaan vai absoluuttisen}

\item{means}{piirretäänkö keskiarvoviivoja ja millä arvoilla}
}
\description{
Tulostaa kuvion, jossa vertailtu suomen ja venäjän prosenttiosuuksia
}
