% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{PrintCompTable}
\alias{PrintCompTable}
\title{Tulostaa side-by-side-tyyppisen taulukon}
\usage{
PrintCompTable(tabs, cnames, gheadings, cap)
}
\arguments{
\item{tabs}{lista vierekkäin tulevista taulukoista}

\item{cnames}{sarakkeiden nimet}

\item{gheadings}{nimetty vektori, joka kertoo myös yhden taulukon sarakkeiden määrän}

\item{cap}{Taulukon otsikko}
}
\description{
Taulukossa yhtä monta saraketta suomessa ja venäjässä
}
