% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{PrintGroupTable}
\alias{PrintGroupTable}
\title{Oikopolku ryhmien kuvausten tulostamiseen taulukoina}
\usage{
PrintGroupTable(g, fi, ru)
}
\description{
Oikopolku ryhmien kuvausten tulostamiseen taulukoina
}
