% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{PrintLocationTableWithCounts}
\alias{PrintLocationTableWithCounts}
\title{Tulostaa taulukon, jossa kielikohtaisesti sijainnit (n + pros)}
\usage{
PrintLocationTableWithCounts(sourcetibble, col1, col2, cap)
}
\arguments{
\item{sourcetibble}{varsinainen data}

\item{col1}{ensimmäisen sarakkeen nimi}

\item{col2}{toisen sarakkeen nimi}

\item{cap}{taulukon otsikko}
}
\description{
Tulostaa taulukon, jossa kielikohtaisesti sijainnit (n + pros)
}
