% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{PrintSampleTable}
\alias{PrintSampleTable}
\title{Oikopolku satunnaisotantataulukoihin}
\usage{
PrintSampleTable(input, cap)
}
\arguments{
\item{input}{syötteenä oleva tibble}

\item{cap}{taulukon otsikko}
}
\description{
Oikopolku satunnaisotantataulukoihin
}
