% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pragtab.R
\docType{methods}
\name{PrintTab,PragmaticAssertionTab-method}
\alias{PrintTab,PragmaticAssertionTab-method}
\title{Prints out a table representing pragmatic assertions/presuppositions}
\usage{
\S4method{PrintTab}{PragmaticAssertionTab}(object)
}
\description{
Prints out a table representing pragmatic assertions/presuppositions
}
