% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaderfunctions.R
\name{ProcessJson}
\alias{ProcessJson}
\title{Lataa dataframe kustakin aineistoryhmästä
Json-tiedoston perusteella ja tarkista, onko tässä ryhmässä
ylimääräisiä tai puuttuvia sarakkeita}
\usage{
ProcessJson(path, lang)
}
\description{
Lataa dataframe kustakin aineistoryhmästä
Json-tiedoston perusteella ja tarkista, onko tässä ryhmässä
ylimääräisiä tai puuttuvia sarakkeita
}
