% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurelists.R
\name{RemoveAdvFeat}
\alias{RemoveAdvFeat}
\alias{RemoveAdvFeat}
\title{Poista piirresarjasta ajan adverbin piirteet ja lyhennä kattamaan vain ympäröivä lause}
\usage{
RemoveAdvFeat(s, tokenid, addedmargin, substractedmargin, sentid = NULL)

RemoveAdvFeat(s, tokenid, addedmargin, substractedmargin, sentid = NULL)
}
\arguments{
\item{s}{kokonainen, muokkaamaton piirresarja}

\item{tokenid}{kuinka mones sana ajanilmauksen ensimmäinen sana on}

\item{addedmargin}{kuinka monta sanaa ajanilmaukseen kuuluu ensimmäisen sanan jälkeen}

\item{substractedmargin}{kuinka monta sanaa ajanilmaukseen kuuluu ennen ensimmäistä sanaa}

\item{sentid}{lauseen id (debugging only)}

\item{s}{kokonainen, muokkaamaton piirresarja}

\item{tokenid}{kuinka mones sana ajanilmauksen ensimmäinen sana on}

\item{addedmargin}{kuinka monta sanaa ajanilmaukseen kuuluu ensimmäisen sanan jälkeen}

\item{substractedmargin}{kuinka monta sanaa ajanilmaukseen kuuluu ennen ensimmäistä sanaa}

\item{sentid}{lauseen id (debugging only)}
}
\description{
Piirresarjalla tarkoitetaan virkkeen sanojen morfologisia ominaisuuksia, jotka on kirjoitettu jonoksi hakasulkeiden väliin. Tämä funktio poistaa ajanilmauksen piirteet piirresarjasta ja pyrkii myös rajaamaan käsiteltävän alueen vain ajanilmauksen sisältäväksi lauseeksi välimerkkien perusteella.

Piirresarjalla tarkoitetaan virkkeen sanojen morfologisia ominaisuuksia, jotka on kirjoitettu jonoksi hakasulkeiden väliin. Tämä funktio poistaa ajanilmauksen piirteet piirresarjasta ja pyrkii myös rajaamaan käsiteltävän alueen vain ajanilmauksen sisältäväksi lauseeksi välimerkkien perusteella.
}
