% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collocations.R
\name{SimpleCollocations}
\alias{SimpleCollocations}
\title{Hae yksinkertainen kollokaattilista}
\usage{
SimpleCollocations(sourcedata, targetpattern, maxcontext = 1)
}
\arguments{
\item{sourcedata}{lista (vektori) lauseista/konteksteista, joista kollokaatteja etsitään}

\item{targetpattern}{se merkkijono(REGEX), jolle kollokaatteja etsitään}

\item{maxcontext}{miten pitkälle kollokaatteja kartotetaan}
}
\description{
Hae yksinkertainen kollokaattilista
}
