% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sov}
\alias{sov}
\title{Venäjän ja suomen SOV-lauseita koskeva numeerinen data.}
\format{tibble seuraavalla rakenteella
\describe{
  \item{lang}{lauseen kieli}
  \item{group}{esimerkin aineistoryhmä}
  \item{funct}{semanttinen funktio}
  \item{morph}{morfologinen rakenne}
  \item{location3}{kolmijakoinen sijaintimuuttuja (S1, S2/S3, S4)}
  \item{location}{nelijakoinen sijaintimuuttuja (S1,S2,S3,S4)}
  \item{isnumeric}{onko lauseessa läsnä numeraali - tarkemmin ks. funktio DoesThisRowHaveNumeral}
  \item{pos}{positionaalisuus}
  \item{red}{referentiaalisuus}
  \item{corpustype}{lähdekorpuksen tyyppi}
  \item{sentid}{lauseen alkuperäinen id tietokannassa}
  ...
}}
\usage{
sov
}
\description{
Venäjän ja suomen SOV-lauseita koskeva numeerinen data.
}
\keyword{datasets}
