% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving_data.R
\name{GetS1Means}
\alias{GetS1Means}
\title{Laske s1-sijainnin keskiarvoja. Kerää myös functtabs-taulukko eri semanttisten funktioiden jakautumisesta eri sijainteihin}
\usage{
GetS1Means()
}
\description{
Laske s1-sijainnin keskiarvoja. Kerää myös functtabs-taulukko eri semanttisten funktioiden jakautumisesta eri sijainteihin
}
